feat: configure UIO integration, enable testing of UIO integration
