+++ .eleventy.js
@@ -1,61 +1,16 @@
-/*
-Copyright the eleventy-plugin-fluid copyright holders.
-
-See the AUTHORS.md file at the top-level directory of this distribution and at
-https://github.com/fluid-project/eleventy-plugin-fluid/raw/main/AUTHORS.md.
-
-Licensed under the New BSD license. You may not use this file except in compliance with this License.
-
-You may obtain a copy of the New BSD License at
-https://github.com/fluid-project/eleventy-plugin-fluid/raw/main/LICENSE.md.
-*/
 "use strict";
 
-const figureShortcode = require("./src/shortcodes/figure-shortcode.js");
-const formatDateFilter = require("./src/filters/format-date-filter.js");
-const isoDateFilter = require("./src/filters/iso-date-filter.js");
-const limitFilter = require("./src/filters/limit-filter.js");
-const markdownFilter = require("./src/filters/markdown-filter.js");
-const slugFilter = require("./src/filters/slug-filter.js");
-const splitFilter = require("./src/filters/split-filter.js");
-const uioShortcodes = require("./src/shortcodes/uio.js");
-const uioAssets = require("./src/config/uio-assets.json");
+const fluidPlugin = require('./index.js');
 
-module.exports = {
-    initArguments: {},
-	configFunction: function(eleventyConfig, options = {}) {
-        options = Object.assign({
-            uio: true,
-        }, options);
+module.exports = function(eleventyConfig) {
+    eleventyConfig.addPlugin(fluidPlugin);
 
-        eleventyConfig.addFilter("formatDate", formatDateFilter);
-        eleventyConfig.addFilter("isoDate", isoDateFilter);
-        eleventyConfig.addFilter("limit", limitFilter);
-        eleventyConfig.addFilter("markdown", markdownFilter);
-        eleventyConfig.addFilter("slug", slugFilter);
-        eleventyConfig.addFilter("split", splitFilter);
-
-        eleventyConfig.addShortcode("figure", figureShortcode);
-
-        if (options.uio) {
-            Object.keys(uioShortcodes).forEach((shortcodeName) => {
-                eleventyConfig.addShortcode(shortcodeName, uioShortcodes[shortcodeName])
-            });
-    
-            uioAssets.forEach((asset) => {
-                const fileMapping = {};
-                fileMapping[`node_modules/infusion/${asset}`] = `lib/infusion/${asset.replace("dist/", "")}`;
-                eleventyConfig.addPassthroughCopy(fileMapping);
-            });
-        }
-
-        return {
-            dir: {
-                input: 'src',
-                output: 'dist',
-                includes: "_includes"
-            },
-            passthroughFileCopy: true
-        }
-    }
+    return {
+        dir: {
+            input: "src",
+            output: "dist",
+            includes: "_includes"
+        },
+        passthroughFileCopy: true
+    };
 };
