+++ .eleventyignore
@@ -1,1 +1,5 @@
 src/**/*.md
+AUTHORS.md
+CHANGELOG.md
+LICENSE.md
+README.md
