+++ .gitignore
@@ -1,2 +1,3 @@
+_site
 dist
 node_modules
