+++ package.json
@@ -5,10 +5,11 @@
     "publishConfig": {
         "access": "public"
     },
-    "main": ".eleventy.js",
+    "main": "index.js",
     "scripts": {
         "build": "eleventy",
         "lint": "grunt lint",
+        "start": "eleventy --serve",
         "test": "ava",
         "release": "standard-version -t ''"
     },
@@ -31,6 +32,7 @@
         }
     },
     "devDependencies": {
+        "@11ty/eleventy": "0.11.1",
         "@commitlint/cli": "11.0.0",
         "@commitlint/config-conventional": "11.0.0",
         "ava": "3.13.0",
@@ -40,7 +42,6 @@
         "standard-version": "9.0.0"
     },
     "dependencies": {
-        "@11ty/eleventy": "0.11.1",
         "infusion": "3.0.0-dev.20201113T153152Z.32176dcbe.FLUID-6145",
         "markdown-it": "12.0.2",
         "slugify": "1.4.6"
