+++ src/shortcodes/uio.js
@@ -40,14 +40,24 @@ module.exports = {
     },
     uioInit: function () {
         return `<script>
-    fluid.uiOptions.prefsEditor(".flc-prefsEditor-separatedPanel", {
-      terms: {
-        "templatePrefix": "/lib/infusion/src/framework/preferences/html",
-        "messagePrefix": "/lib/infusion/src/framework/preferences/messages"
-      },
-      "tocTemplate": "/lib/infusion/src/components/tableOfContents/html/TableOfContents.html",
-      "tocMessage": "/lib/infusion/src/framework/preferences/messages/tableOfContents-enactor.json"
-    });
-  </script>`;
+            fluid.uiOptions(".flc-prefsEditor-separatedPanel", {
+                auxiliarySchema: {
+                    terms: {
+                        "templatePrefix": "/lib/infusion/src/framework/preferences/html",
+                        "messagePrefix": "/lib/infusion/src/framework/preferences/messages"
+                    },
+                    "fluid.prefs.tableOfContents": {
+                        enactor: {
+                            "tocTemplate": "/lib/infusion/src/components/tableOfContents/html/TableOfContents.html",
+                            "tocMessage": "/lib/infusion/src/framework/preferences/messages/tableOfContents-enactor.json",
+                            ignoreForToC: {}
+                        }
+                    }
+                },
+                prefsEditorLoader: {
+                    lazyLoad: true
+                }
+            });
+        </script>`;
     }
 };
